+++ src/main/java/insane96mcp/enhancedai/mixin/WitchMixin.java
@@ -1,12 +1,21 @@
 package insane96mcp.enhancedai.mixin;
 
+import com.llamalad7.mixinextras.expression.Definition;
+import com.llamalad7.mixinextras.expression.Expression;
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.llamalad7.mixinextras.sugar.Local;
+import insane96mcp.enhancedai.data.PotionOrMobEffect;
 import insane96mcp.enhancedai.modules.witch.ThirstyWitches;
 import insane96mcp.enhancedai.modules.witch.darkart.DarkArt;
 import insane96mcp.insanelib.base.Feature;
+import insane96mcp.insanelib.util.MCUtils;
 import insane96mcp.insanelib.util.ModNBTData;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.tags.DamageTypeTags;
+import net.minecraft.world.effect.MobEffect;
+import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
@@ -16,7 +25,12 @@ import net.minecraft.world.entity.monster.Witch;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.raid.Raider;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.alchemy.Potion;
+import net.minecraft.world.item.alchemy.PotionUtils;
+import net.minecraft.world.item.alchemy.Potions;
 import net.minecraft.world.level.Level;
+import net.minecraftforge.common.ForgeMod;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
@@ -42,7 +56,7 @@ public abstract class WitchMixin extends Raider {
 	private int usingTime;
 
 	@Unique
-    private int enhancedAI$invisibilityCooldown = 20;
+    private int enhancedai$invisibilityCooldown = 20;
 
 	protected WitchMixin(EntityType<? extends Raider> p_37839_, Level p_37840_) {
 		super(p_37839_, p_37840_);
@@ -62,35 +76,77 @@ public abstract class WitchMixin extends Raider {
 		return alive && !ModNBTData.get(this, DarkArt.PERFORMING_DARK_ARTS, Boolean.class);
 	}
 
-	@WrapOperation(method = "aiStep", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/monster/Witch;setItemSlot(Lnet/minecraft/world/entity/EquipmentSlot;Lnet/minecraft/world/item/ItemStack;)V"))
+	@ModifyExpressionValue(method = "aiStep", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/monster/Witch;isDrinkingPotion()Z"))
+	public boolean enhancedai$playSoundWhenDrinking(boolean original) {
+		if (original && this.usingTime % 8 == 0)
+			this.playSound(SoundEvents.GENERIC_DRINK, 1.0f, this.random.nextFloat() * 0.1F + 0.9F);
+		return original;
+	}
+
+	@Unique
+	private ItemStack enhancedai$stackToUse;
+
+	@Definition(id = "potion", local = @Local(type = Potion.class))
+	@Expression("potion != null")
+	@ModifyExpressionValue(method = "aiStep", at = @At("MIXINEXTRAS:EXPRESSION"))
+	private boolean enhancedai$newUseItem(boolean original) {
+		if (!Feature.isEnabled(ThirstyWitches.class))
+			return original;
+
+		enhancedai$stackToUse = ItemStack.EMPTY;
+		double distance = ThirstyWitches.CUSTOM_DRINK_DISTANCE_SAFE.get(this);
+		distance *= distance;
+		if (this.getTarget() != null && this.getTarget() instanceof Player && this.distanceToSqr(this.getTarget()) > distance) {
+			for (PotionOrMobEffect potionOrMobEffect : ThirstyWitches.drinkPotion) {
+				MobEffect mobEffect = potionOrMobEffect.getMobEffect();
+				if (mobEffect != null && this.hasEffect(mobEffect))
+					continue;
+
+				enhancedai$stackToUse = potionOrMobEffect.getPotionStack();
+				break;
+			}
+		}
+		else {
+			Potion potion = null;
+			if (this.random.nextFloat() < ThirstyWitches.WATER_BREATHING_CHANCE.get(this)
+					&& this.isEyeInFluidType(ForgeMod.WATER_TYPE.get())
+					&& !this.hasEffect(MobEffects.WATER_BREATHING)
+					&& this.getAirSupply() < this.getMaxAirSupply() / 2)
+				potion = Potions.WATER_BREATHING;
+			else if (this.random.nextFloat() < ThirstyWitches.FIRE_RESISTANCE_CHANCE.get(this)
+					&& (this.isOnFire() || this.getLastDamageSource() != null && this.getLastDamageSource().is(DamageTypeTags.IS_FIRE))
+					&& !this.hasEffect(MobEffects.FIRE_RESISTANCE))
+				potion = Potions.FIRE_RESISTANCE;
+			else if (this.getHealth() / this.getMaxHealth() < ThirstyWitches.HEALING_THRESHOLD.get(this)
+					&& this.random.nextFloat() < ThirstyWitches.HEALING_CHANCE.get(this)) {
+				potion = Potions.HEALING;
+				if (this.getHealth() / this.getMaxHealth() < ThirstyWitches.STRONG_HEALING_THRESHOLD.get(this))
+					potion = Potions.STRONG_HEALING;
+			}
+			if (potion != null)
+				enhancedai$stackToUse = PotionUtils.setPotion(new ItemStack(Items.POTION), potion);
+		}
+
+		if (enhancedai$stackToUse == null && MCUtils.hasLongNegativeEffect(this) && this.random.nextDouble() < ThirstyWitches.MILK_CHANCE.get(this))
+			enhancedai$stackToUse = new ItemStack(Items.MILK_BUCKET);
+		return !enhancedai$stackToUse.isEmpty();
+	}
+
+	@WrapOperation(method = "aiStep", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/monster/Witch;setItemSlot(Lnet/minecraft/world/entity/EquipmentSlot;Lnet/minecraft/world/item/ItemStack;)V", ordinal = 1))
 	private void enhancedai$changeUseItem(Witch instance, EquipmentSlot equipmentSlot, ItemStack itemStack, Operation<Void> original) {
 		if (!Feature.isEnabled(ThirstyWitches.class)) {
 			original.call(instance, equipmentSlot, itemStack);
 			return;
 		}
 
-
+		original.call(instance, equipmentSlot, enhancedai$stackToUse);
 	}
 
 	@Inject(at = @At("HEAD"), method = "aiStep", cancellable = true)
 	private void aiStep(CallbackInfo ci) {
-        /*if (!Modules.witch.isEnabled())
-            return;
-
-        ci.cancel();
-
-        if (this.level().isClientSide
-				|| !this.isAlive()) {
-			super.aiStep();
-			return;
-		}
-
-		this.healRaidersGoal.decrementCooldown();
-		this.attackPlayersGoal.setCanAttack(this.healRaidersGoal.getCooldown() <= 0);
-
+/*
 		if (this.isDrinkingPotion()) {
-			if (this.usingTime % 8 == 0)
-				this.playSound(SoundEvents.GENERIC_DRINK, 1.0f, this.random.nextFloat() * 0.1F + 0.9F);
+
 			if (this.usingTime-- <= 0) {
 				this.setUsingItem(false);
 				ItemStack itemstack = this.getMainHandItem();