+++ src/main/java/insane96mcp/enhancedai/modules/witch/darkart/DarkArt.java
@@ -20,7 +20,7 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 @LoadFeature(module = Modules.Ids.WITCH, description = "Witches summon Villagers and cast a lightning upon them. Only entity types in the `enhancedai:witch/can_perform_dark_arts` tag will be affected (can be used for any mob, not only witches).")
 public class DarkArt extends Feature {
-	public static final TagKey<EntityType<?>> CAN_PERFORM = TagKey.create(Registries.ENTITY_TYPE, EnhancedAI.location("witch/can_perform_dark_arts"));
+	public static final TagKey<EntityType<?>> AFFECTED_ENTITY_TYPES = TagKey.create(Registries.ENTITY_TYPE, EnhancedAI.location("witch/can_perform_dark_arts"));
     @Config(min = 0d, max = 1d, description = "Chance for a witch to get the Dark Art AI (as soon as they have a target and are less than 10 blocks away from the target will summon a Villager and cast a lightning bolt on them")
     public static Double chance = 0.333d;
 	@Config(min = 0d, description = "At which distance from the witch will she start performing the Dark Art")
@@ -49,7 +49,7 @@ public class DarkArt extends Feature {
         if (!this.isEnabled()
                 || event.getLevel().isClientSide
 				|| !(event.getEntity() instanceof Mob mob)
-                || !mob.getType().is(CAN_PERFORM))
+                || !mob.getType().is(AFFECTED_ENTITY_TYPES))
             return;
 
 		DARK_ARTS.applyIfAbsent(mob, mob.getRandom().nextDouble() < chance);